Extract database backup snapshots
